cart || admin auth || create order

+++ src/controller/cart/cart.controller.js
@@ -1,7 +1,10 @@
-import path from "path";
+import crypto from 'crypto';
 import { sendError, sendSuccess } from "../../../utils/resHandler.js";
 import Cart from "../../model/Cart/cart.model.js";
 import CartIteam from "../../model/Cart/cartItem.model.js";
+import razorpay from "../../../utils/razorpay.js";
+import Order from '../../model/Cart/order.model.js';
+import mongoose from 'mongoose';
 
 const addCardItem = async (req,res)=>{
     const {subcategoryId,qty,total} = req.body;
@@ -25,10 +28,11 @@ const addCardItem = async (req,res)=>{
         const cart =  await Cart.create({
           userId,
           Iteam:[cartItemcreate._id],
-          subtotal:cartItemcreate.total
+          subtotal:cartItemcreate.total,
+          deliveryFee:50
         });
         console.log('the cart is',cart)
-        sendSuccess(res,"the cart is created successfully",201)
+        return  sendSuccess(res,"the cart is created successfully",201)
       } 
 
        const existingCartIteam = cartdata.Iteam.find(item => item.subcategoryId.toString() === subcategoryId);
@@ -41,7 +45,7 @@ const addCardItem = async (req,res)=>{
          await existingCartIteam.save();
          await cartdata.save();
 
-         sendSuccess(res,existingCartIteam,"the cartIteam added successfully",201)
+        return sendSuccess(res,existingCartIteam,"the cartIteam added successfully",201)
        }
        
        const newCartIteam = await CartIteam.create({
@@ -54,7 +58,7 @@ const addCardItem = async (req,res)=>{
        cartdata.Iteam.push(newCartIteam._id)
        cartdata.subtotal+=total
        await cartdata.save();
-       return sendSuccess(res, cartdata, 200);
+       sendSuccess(res, cartdata, 200);
 
            
       }catch(err){
@@ -84,13 +88,10 @@ const showallCartdata = async (req,res) => {
           }
           // const totalprice = await Cart.find(userId)
             
-         const subtotalprice =  cartDataforuser.reduce((acc,currentval)=>{
-                console.log('the currentval is',currentval);
-               return acc + (currentval.total || 0)
-          },0);
-          console.log('the subtotalprice price',subtotalprice);
+        
+          
 
-          sendSuccess(res,{cartDataforuser,subtotalprice},200);
+          sendSuccess(res,cartDataforuser,200);
          
     }catch(err){
       console.log('the error occur in the cartdataforUser',err.message)
@@ -100,13 +101,30 @@ const showallCartdata = async (req,res) => {
 const cartCategorydelete = async (req,res)=>{
         const {deleteId} = req.params;
         console.log('deleteID is',deleteId)
-        try{
-            const deleteData = await CartIteam.findByIdAndDelete(deleteId);
-            console.log('the deleteData is',deleteData);
 
-            if(!deleteData){
-              return sendError(res,"item is not found",401)
+        if(!mongoose.Types.ObjectId.isValid(deleteId)) return sendError(res,'invalid item id',400)
+        try{
+            const cart = await Cart.findOne({userId:req.user.sub}).populate('Iteam');
+            console.log("this is cart for the deletd item",cart)
+            if(!cart){
+              return sendError(res,"cart is not found",404)
             }
+
+            const iteamtoDelete = cart.Iteam.find(item => item._id.toString() === deleteId);
+            console.log('the deletedData is',iteamtoDelete)
+            if(!iteamtoDelete) {
+              return sendError(res,"item is not found in cart",404)
+            }
+
+            await CartIteam.findByIdAndDelete(deleteId);
+            const fillterval = cart.Iteam.filter((item)=> item._id.toString() !== deleteId);
+            console.log('fillter value is',fillterval)
+            cart.Iteam = fillterval
+            cart.subtotal -= iteamtoDelete.total
+            cart.save();
+            // console.log('the deleteData is',deleteData);
+            
+            
             sendSuccess(res,"item is deleted successfull",200);
         }catch(err){
              console.log('the error occur in the cartCategorydelete',err.message)
@@ -114,11 +132,95 @@ const cartCategorydelete = async (req,res)=>{
         }
 }
 
+const razorpayOrderid = async (req,res)=>{
+       const userId = req.user.sub;
+
+       try{
+            const cartData = await Cart.findOne({userId}).populate('Iteam');
+            console.log('the cartData',cartData);
+            const options = {
+              amount:cartData.grandtotal*100,
+              currency:"INR",
+              receipt:"order_rcptid_11"
+            }
+
+           const order =await razorpay.orders.create(options);
+           console.log('the order is provide',order)
+
+           sendSuccess(res,{orderId:order.id,key:process.env.RAZORPAY_ID,amount:order.amount})
+
+       }catch(err){
+        console.log("the errror in the razorpay for order create",err.message)
+       }
+
+}
+
+const checkOut = async (req,res)=>{
+   const {address,paymentDetail} = req.body;
+   console.log('the req.body is',req.body);
+
+   try{
+       
+      const cart = await Cart.findOne({userId:req.user.sub}).populate('Iteam');
+      
+      console.log('the cart is',cart);
+
+      if(!cart || cart.Iteam.length === 0) return sendError(res,"Cart is empty",404)
+
+        const deliveryFee = 50;
+        const grandTotal = cart.subtotal + deliveryFee;  
+        
+        const {razorpay_order_id,razorpay_payment_id,razorpay_signature} = paymentDetail;
+        console.log('razorpay_signature',razorpay_signature);
+        const generatedSignature = crypto.
+                                  createHmac('sha256',process.env.RAZORPAY_SECRET)
+                                  .update(razorpay_order_id +"|"+ razorpay_payment_id)
+                                  .digest('hex');
+        console.log('generatedSignature',generatedSignature);
+
+        if(generatedSignature !== razorpay_signature){
+           return sendError(res,"payment are not successfully happen ")
+        }
+   
+        const ordercreate = await Order.create({
+           userId:req.user.sub,
+           items:cart.Iteam.map(item => item._id),
+           grandTotal:grandTotal,
+           deliveryAddress:address,
+           paymentStatus:'success'
+        })
+
+        console.log('the order is create',ordercreate)
+        
+         sendSuccess(res,'order is created successfully',201)
+      
+   } catch(err){
+     console.log('the error occur in checkOut',err.message)
+   }
+}
+
+const userOrder =  async (req,res) => {
+     try{
+         
+         const userOrder = await Order.findOne({userId:req.user.sub}).populate('items');
+
+         if(!userOrder) return sendError(res,'the user have not any order',404)
+
+        sendSuccess(res,userOrder,201);
+
+        
+     }catch(err){
+      console.log('the error in the userOrder',err.message)
+     }
+}
 
 export {
     
     showallCartdata,
     cartCategorydelete,
-    addCardItem
+    addCardItem,
+    razorpayOrderid,
+    checkOut,
+    userOrder
 };
